some fix

+++ blog-admin/src/layout/ArticleList.jsx
@@ -1,66 +1,15 @@
 import React, {Component, Fragment} from 'react';
-import {deleteArticle, getArticles, likeFuzzyTitle, likeTitlePrefix} from "service/article.js";
+import {deleteArticle, likeFuzzyTitle, likeTitlePrefix, search} from "service/article.js";
 import {AutoComplete, Card, Icon, Input, List, message, Modal, Select, Tooltip} from "antd";
 import css from "./ArticleList.less";
 import commonCss from "common/common.less";
-import {ButtonLink} from "ui";
 import _ from "lodash";
 import {ArticleModal} from "component/ArticleModal.jsx";
 import {ModalWrapper} from "ui/Modal.jsx";
 import {NavLink, withRouter} from "react-router-dom";
 import {pickQuery, searchAssign} from "util/router.js";
 import {pageableToPagination} from "ui/Pager.jsx";
-
-const ArticleItem = ({item, onEdit, onView, onDeny, onDelete}) => {
-    let {id, title, created, modified, views, comments} = item;
-    return (
-        <div key={id} className={css.article}>
-            <h2 onClick={onEdit}><NavLink to={"/article/" + id}>{title}</NavLink></h2>
-            <div className={commonCss.flexBox}>
-                <div>
-                    <span className={commonCss.span}>
-                        <Icon type="eye"/>
-                        <span className={css.num}>{views}</span>
-                    </span>
-                    <span className={commonCss.span}>
-                        <Icon type="message"/>
-                        <span className={css.num}>{comments}</span>
-                    </span>
-                    <span className={commonCss.span}>创建时间：{created}</span>
-                    <span className={commonCss.span}>最后修改时间：{modified}</span>
-                </div>
-                <div className={commonCss.flex1}/>
-                <div>
-                    <ButtonLink className={commonCss.span} onClick={() => onView(id)}>查看</ButtonLink>
-                    <ButtonLink className={commonCss.span} onClick={() => onDeny(id)}>禁止评论</ButtonLink>
-                    <ButtonLink className={commonCss.span} onClick={() => onDelete(id)}>删除</ButtonLink>
-                </div>
-            </div>
-        </div>
-    );
-};
-
-const SearchForm = ({dataSource, onCompleteSearch, onSearch}) => {
-    let options = _.map(dataSource, ({id, title}) => (
-        <Select.Option key={id} value={_.toString(id)}>{title}</Select.Option>)
-    );
-    return (
-        <AutoComplete dataSource={options} onSearch={onCompleteSearch}>
-            <Input.Search placeholder="输入文章标题" onSearch={onSearch}/>
-        </AutoComplete>
-    );
-};
-
-const CardExtra = ({searchSuggest, onNew, onCompleteSearch, onSearch}) => (
-    <Fragment>
-        <Tooltip title="新建">
-            <Icon className={css.edit} type="edit" onClick={onNew}/>
-        </Tooltip>
-        <SearchForm dataSource={searchSuggest}
-                    onCompleteSearch={onCompleteSearch}
-                    onSearch={onSearch}/>
-    </Fragment>
-);
+import {ButtonLink} from "ui";
 
 @withRouter
 export class ArticleList extends Component {
@@ -79,8 +28,13 @@ export class ArticleList extends Component {
 
     async fetchData(published, page, size) {
         this.setState({published, page, size, loading: true});
-        const data = await getArticles(published, page, size);
+        const data = await search({published, deleted: false}, page, size);
         this.setState({data, loading: false});
+    }
+
+    refresh() {
+        const {published, page, size} = this.state;
+        this.fetchData(published, page, size);
     }
 
     handleTabChange = (published) => searchAssign({published, page: 1});
@@ -98,17 +52,22 @@ export class ArticleList extends Component {
     };
 
     handleDelete = (articleId) => {
-        Modal.warning({
+        Modal.confirm({
+            title: "确认删除改文章",
             content: "确认删除改文章",
             onOk: async () => {
                 try {
                     await deleteArticle(articleId);
                     message.success("删除成功");
+                    this.refresh();
                 } catch (e) {
                     console.log(e)
                 }
             }
         });
+    };
+    handleDeny = (articleId) => {
+        console.log(articleId);
     };
     handleNew = () => {
         this.setState({showModal: true});
@@ -118,19 +77,62 @@ export class ArticleList extends Component {
         this.setState({showModal: false});
     };
 
+    renderCardExtra = () => {
+        const {searchSuggest = []} = this.state;
+        const options = _.map(searchSuggest, ({id, title}) => (
+            <Select.Option key={id} value={_.toString(id)}>{title}</Select.Option>)
+        );
+        return (
+            <Fragment>
+                <Tooltip title="新建">
+                    <Icon className={css.edit} type="edit" onClick={this.handleNew}/>
+                </Tooltip>
+                <AutoComplete dataSource={options} onSearch={this.handleCompleteSearch}>
+                    <Input.Search placeholder="输入文章标题" onSearch={this.handleSearch}/>
+                </AutoComplete>
+            </Fragment>
+        );
+    };
+
+    renderItem = (item) => {
+        let {id, title, created, modified, views, comments} = item;
+        return (
+            <div key={id} className={css.article}>
+                <h2><NavLink to={`/article/${id}`}>{title}</NavLink></h2>
+                <div className={commonCss.flexBox}>
+                    <div>
+                    <span className={commonCss.span}>
+                        <Icon type="eye"/>
+                        <span className={css.num}>{views}</span>
+                    </span>
+                        <span className={commonCss.span}>
+                        <Icon type="message"/>
+                        <span className={css.num}>{comments}</span>
+                    </span>
+                        <span className={commonCss.span}>创建时间：{created}</span>
+                        <span className={commonCss.span}>最后修改时间：{modified}</span>
+                    </div>
+                    <div className={commonCss.flex1}/>
+                    <div>
+                        <ButtonLink href={`/article/${id}`}>查看</ButtonLink>
+                        <ButtonLink onClick={() => this.handleDeny(id)}>禁止评论</ButtonLink>
+                        <ButtonLink onClick={() => this.handleDelete(id)}>删除</ButtonLink>
+                    </div>
+                </div>
+            </div>
+        );
+    };
+
     render() {
-        const {data = {}, searchSuggest = []} = this.state;
+        const {loading, published, data = {}} = this.state;
         const {content, totalElements, pageable} = data;
         return (
-            <Card title="文章列表" activeTabKey={_.toString(this.state.published)}
+            <Card title="文章列表" activeTabKey={_.toString(published)}
                   tabList={this.tabList} onTabChange={this.handleTabChange}
-                  extra={<CardExtra searchSuggest={searchSuggest}
-                                    onSearch={this.handleSearch} onNew={this.handleNew}
-                                    onCompleteSearch={this.handleCompleteSearch}/>}>
+                  extra={this.renderCardExtra()}>
                 <List dataSource={content}
-                      loading={this.state.loading}
-                      renderItem={item => <ArticleItem key={item.id} item={item}
-                                                       onDelete={this.handleDelete}/>}
+                      loading={loading}
+                      renderItem={this.renderItem}
                       pagination={{
                           showSizeChanger: true,
                           showQuickJumper: true,
